Move make/take message back to utility
